+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+- 0.3.2
+    - Refactor single sample inference methods
+    - Update the docs site
+
 - 0.3.1
     - Fix CART feature importances purity increase overflow